% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cad.R
\docType{data}
\name{cad}
\alias{cad}
\title{Knowledge spaces on AutoCAD knowledge}
\format{
A list containing seven bases (cad1 to cad6, and cadmaj) in binary
matrix form. Each matrix has 28 columns representing the different knowledge
items and a varying number of rows containing the basis elements.
}
\usage{
cad
}
\description{
Bases of knowledge spaces on AutoCAD knowledge obtained from querying experts.
}
\details{
Six experts were queried about prerequisite relationships between 28 AutoCAD
knowledge items (Dowling, 1991; 1993). A seventh basis represents those
prerequisite relationships on which the majority (4 out of 6) of the experts
agree (Dowling & Hockemeyer, 1998).
}
\references{
Dowling, C. E. (1991). Constructing Knowledge Structures from
the Judgements of Experts. Habilitationsschrift, Technische Universität
Carolo-Wilhelmina, Braunschweig, Germany.

Dowling, C. E. (1993). Applying the basis of a knowledge space for controlling
the questioning of an expert. Journal of Mathematical Psychology, 37, 21–48.

Dowling, C. E. & Hockemeyer, C. (1998). Computing the intersection of knowledge
spaces using only their basis. In Cornelia E. Dowling, Fred S. Roberts, & Peter
Theuns, editors, Recent Progress in Mathematical Psychology, pp. 133–141.
Lawrence Erlbaum Associates Ltd., Mahwah, NJ.
}
\keyword{data}
