% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fractions.R
\docType{data}
\name{fractions}
\alias{fractions}
\title{Knowledge spaces on fractions}
\format{
A list containing four bases (frac1 to frac3, and fracmaj) in binary
matrix form. Each matrix has 77 columns representing the different knowledge
items and a varying number of rows containing the basis elements.
}
\usage{
fractions
}
\description{
Bases of knowledge spaces on fractions obtained from querying experts.
}
\details{
Three experts were queried about prerequisite relationships between 77 items
on fractions (Baumunk & Dowling, 1997). A forth basis represents those
prerequisite relationships on which the majority of the experts agree
(Dowling & Hockemeyer, 1998).
}
\references{
Baumunk, K. & Dowling, C. E. (1997). Validity of spaces for assessing
knowledge about fractions. \emph{Journal of Mathematical Psychology, 41,} 99–105.

Dowling, C. E. & Hockemeyer, C. (1998). Computing the intersection of knowledge
spaces using only their basis. In Cornelia E. Dowling, Fred S. Roberts, & Peter
Theuns, editors, \emph{Recent Progress in Mathematical Psychology,} pp. 133–141.
Lawrence Erlbaum Associates Ltd., Mahwah, NJ.
}
\seealso{
Other Data: 
\code{\link{cad}},
\code{\link{readwrite}},
\code{\link{xpl}}
}
\concept{Data}
\keyword{datasets}
