% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmassesshalfsplit.R
\name{kmassesshalfsplit}
\alias{kmassesshalfsplit}
\title{Determine next question for probabilistic knowledge assessment}
\usage{
kmassesshalfsplit(probs, ks)
}
\arguments{
\item{probs}{Probability distribution over the knowledge structure (vector)}

\item{ks}{Binary matrix of the knowledge structure}
}
\value{
Number of the selected question
}
\description{
\code{kmassesshalfsplit} determines the next question in a probabilistic
assessment according to the halfsplit rule.
}
\examples{
kmassesshalfsplit(c(0.02, 0.1, 0.07, 0.01, 0.4, 0.17, 0.07, 0.08, 0.08),
                  xpl$space)

}
\seealso{
Other Knowledge assessment: 
\code{\link{kmassess}()},
\code{\link{kmassessbayesian}()},
\code{\link{kmassessinformative}()},
\code{\link{kmassessmentsimulation}()},
\code{\link{kmassessmultiplicative}()},
\code{\link{kmsassess}()}
}
\concept{Knowledge assessment}
