% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmbasis.kmsurmisefunction.R
\name{kmbasis.kmsurmisefunction}
\alias{kmbasis.kmsurmisefunction}
\title{Determine the basis for a surmise function}
\usage{
\method{kmbasis}{kmsurmisefunction}(x, ...)
}
\arguments{
\item{x}{Surmise Function}

\item{...}{Optional/future parameters}
}
\value{
Basis
}
\description{
Determine the basis for a surmise function
}
\seealso{
Other Different representations for knowledge spaces: 
\code{\link{kmSF2basis}()},
\code{\link{kmSR2basis}()},
\code{\link{kmbasis.kmsurmiserelation}()},
\code{\link{kmbasis.matrix}()},
\code{\link{kmsurmisefunction}()},
\code{\link{kmsurmiserelation}()},
\code{\link{kmunionclosure}()}
}
\concept{Different representations for knowledge spaces}
