% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmbasis.kmsurmiserelation.R
\name{kmbasis.kmsurmiserelation}
\alias{kmbasis.kmsurmiserelation}
\title{Determine the basis of a knowledge space from a surmise relation}
\usage{
\method{kmbasis}{kmsurmiserelation}(x, ...)
}
\arguments{
\item{x}{Surmise relation}

\item{...}{Space for additional, optional arameters}
}
\value{
Basis
}
\description{
\code{kmbasis.kmsurmiserelation} takes a surmise relation and returns the
corresponding basis.
}
\seealso{
Other Different representations for knowledge spaces: 
\code{\link{kmSF2basis}()},
\code{\link{kmSR2basis}()},
\code{\link{kmbasis.kmsurmisefunction}()},
\code{\link{kmbasis.matrix}()},
\code{\link{kmsurmisefunction}()},
\code{\link{kmsurmiserelation}()},
\code{\link{kmunionclosure}()}
}
\concept{Different representations for knowledge spaces}
