% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmbasisdiagram.R
\name{kmbasisdiagram}
\alias{kmbasisdiagram}
\title{Plot the Hasse diagram of a basis stored as a matrix}
\usage{
kmbasisdiagram(struc, horizontal = TRUE, colors = NULL)
}
\arguments{
\item{struc}{Binary matrix representing a basis}

\item{horizontal}{Boolean defining orientation of the graph, default TRUE}

\item{colors}{Color vector (default NULL)}
}
\description{
\code{kmbasisdiagram} takes a matrix representing a basis and a
color vector and draws a Hasse diagram. If the color vector is NULL
the states are drawn in green.
}
\seealso{
Other Plotting knowledge structures: 
\code{\link{kmSRdiagram}()},
\code{\link{kmcolors}()},
\code{\link{kmhasse}()}
}
\concept{Plotting knowledge structures}
