% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmcolors.R
\name{kmcolors}
\alias{kmcolors}
\title{Determine a color vector based on probabilities}
\usage{
kmcolors(prob, palette = cm.colors)
}
\arguments{
\item{prob}{Probability vector}

\item{palette}{Color palette (default = cm.colors)}
}
\description{
\code{kmcolors} takes a probabilty vector and a color palette and
creates a color vector to be used with \code{kstMatrix::plot}.
}
\seealso{
Other Plotting knowledge structures: 
\code{\link{plot}()}

Other Utilities: 
\code{\link{kmdoubleequal}()},
\code{\link{kmsetiselement}()},
\code{\link{kmsymmsetdiff}()}
}
\concept{Plotting knowledge structures}
\concept{Utilities}
