% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmdoubleequal.R
\name{kmdoubleequal}
\alias{kmdoubleequal}
\title{Test two double numbers on equity with a certain tolerance}
\usage{
kmdoubleequal(x, y, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{First double to compare}

\item{y}{Second double to compare}

\item{tol}{Tolerance optional)}
}
\value{
Boolean for (approximate) equity
}
\description{
Test two double numbers on equity with a certain tolerance
}
\examples{
kmdoubleequal(0.5+0.5, 1)

}
\seealso{
Other Utilities: 
\code{\link{kmsymmsetdiff}()}
}
\concept{Utilities}
