% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmeqreduction.R
\name{kmeqreduction}
\alias{kmeqreduction}
\title{Reduce a family of knowledge states with respect to item equivalence}
\usage{
kmeqreduction(x)
}
\arguments{
\item{x}{Binary matrix}
}
\value{
Binary matrix reduced by equivalences
}
\description{
\code{kmeqreduction} takes a family of knowledge states and returns its
reduction to non-equivalent items.
}
\examples{
kmeqreduction(xpl$space)

}
\seealso{
Other Properties of knowledge structures: 
\code{\link{kmiswellgraded}()},
\code{\link{kmnotions}()}
}
\concept{Properties of knowledge structures}
