% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmfamset.R
\name{kmfamset}
\alias{kmfamset}
\title{Convert a binary matrix to a \code{kmfamset} object (family of sets)}
\usage{
kmfamset(x)
}
\arguments{
\item{x}{Binary matrix representing a family of sets}
}
\value{
Distance distribution vector
}
\description{
\code{kmfamset} returns a \code{kmfamset} object after checking that
the passed object is a binary matrix with all different rows. If the
passe object inherits the \code{kmfamset} property, nothing else is
changed.
}
\examples{
m <- as.matrix(c(1,0,0,0,1,0,1,1,1), nrow=3, byrow=TRUE)
kmfamset(m)

}
\seealso{
Other Constructors: 
\code{\link{kmspace}()},
\code{\link{kmstructure}()}
}
\concept{Constructors}
