% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmgradations.R
\name{kmgradations}
\alias{kmgradations}
\title{Determine all gradations between two states}
\usage{
kmgradations(structure, from = NULL, to = NULL)
}
\arguments{
\item{structure}{Knowledge structure}

\item{from}{Starting state (if NULL (default), it is the empty set)}

\item{to}{Goal state (if NULL (default), it is the full item set)}
}
\value{
A list of gradations where each gradation is a list of states
}
\description{
Determine all gradations between two states
}
\examples{
kmgradations(xpl$space)


}
\seealso{
Other Fringes & learning paths: 
\code{\link{kmbasisfringe}()},
\code{\link{kmbasisneighbourhood}()},
\code{\link{kmfringe}()},
\code{\link{kmlearningpaths}()},
\code{\link{kmneighbourhood}()},
\code{\link{kmnneighbourhood}()}
}
\concept{Fringes & learning paths}
