% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmiita2SR.R
\name{kmiita2SR}
\alias{kmiita2SR}
\title{Convert an IITA result into a surmise relation matrix}
\usage{
kmiita2SR(ii, names = NULL)
}
\arguments{
\item{ii}{\code{iita()} result}

\item{names}{Vector of item names (default NULL)}
}
\value{
Surmise relation matrix
}
\description{
\code{kmiita2SR} takes the result of a \code{DAKS::iita()} call and
delivers the matrix of the computed surmise relation.
}
\seealso{
Other Generating knowledge spaces: 
\code{\link{kmgenerate}()}
}
\concept{Generating knowledge spaces}
