% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmiswellgraded.R
\name{kmiswellgraded}
\alias{kmiswellgraded}
\title{Check for wellgradedness of a knowledge structure}
\usage{
kmiswellgraded(x)
}
\arguments{
\item{x}{Binary matrix representing a knowledge space}
}
\value{
Logical value specifying whether \code{x} is wellgraded
}
\description{
\code{kmiswellgraded} returns whether a knowledge structure is wellgraded.
}
\examples{
kmiswellgraded(xpl$space)

}
\references{
Doignon, J.-P. & Falmagne, J.-C. (1999). \emph{Knowledge Spaces.}
Springer–Verlag, Berlin.
}
\seealso{
Other Properties of knowledge structures: 
\code{\link{kmeqreduction}()},
\code{\link{kmnotions}()}
}
\concept{Properties of knowledge structures}
