% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmlearningpaths.R
\name{kmlearningpaths}
\alias{kmlearningpaths}
\title{Determine all learning paths in a knowledge structure}
\usage{
kmlearningpaths(structure)
}
\arguments{
\item{structure}{Knowledge structure}
}
\value{
A list of learning paths where each learning path is a
list of states
}
\description{
Determine all learning paths in a knowledge structure
}
\examples{
kmlearningpaths(xpl$space)


}
\seealso{
Other Fringes & learning paths: 
\code{\link{kmbasisfringe}()},
\code{\link{kmbasisneighbourhood}()},
\code{\link{kmfringe}()},
\code{\link{kmgradations}()},
\code{\link{kmneighbourhood}()},
\code{\link{kmnneighbourhood}()}
}
\concept{Fringes & learning paths}
