% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmminimalfamset.R
\name{kmminimalfamset}
\alias{kmminimalfamset}
\title{Determine subfamily of minimal states in a \code{kmfamset}}
\usage{
kmminimalfamset(x)
}
\arguments{
\item{x}{kmfamset}
}
\value{
kmfamset of minimal states
}
\description{
Determine subfamily of minimal states in a \code{kmfamset}
}
\examples{
m <- matrix(c(1,0,0,0,1,0,1,0,1), ncol=3, byrow=TRUE)
m
kmminimalfamset(kmfamset(m))


}
\seealso{
Other Utilities:
\code{\link[=kmcolors]{kmcolors()}},
\code{\link[=kmdoubleequal]{kmdoubleequal()}},
\code{\link[=kmsetiselement]{kmsetiselement()}},
\code{\link[=kmsymmsetdiff]{kmsymmsetdiff()}}
}
\concept{Utilities}
