% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmnotions.R
\name{kmnotions}
\alias{kmnotions}
\title{Determine the notions of a knowledge structure}
\usage{
kmnotions(x)
}
\arguments{
\item{x}{Binary matrix representing a knowledge structure}
}
\value{
Binary matrix representing notions in the knowledge structure

The matrix has a '1' in row 'i' and column 'j' if 'i' and 'j' belong
to the same notion (i.e. are equivalent). It is a symmetric matrix
with '1's in the main diagonal.
}
\description{
\code{kmnotions} returns a matrix representing the
notions of a knowledge structure.
}
\examples{
kmnotions(xpl$space)

}
\seealso{
Other Properties of knowledge structures: 
\code{\link{kmeqreduction}()},
\code{\link{kmiswellgraded}()}
}
\concept{Properties of knowledge structures}
