% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmsetiselement.R
\name{kmsetiselement}
\alias{kmsetiselement}
\title{Test if a state is contained in a family of states}
\usage{
kmsetiselement(s, f)
}
\arguments{
\item{s}{State}

\item{f}{Family of sets}
}
\value{
Boolean is s is contained in f
}
\description{
Test if a state is contained in a family of states
}
\examples{
kmsetiselement(c(1,1,1,0), xpl$space)

}
\seealso{
Other Utilities: 
\code{\link{kmcolors}()},
\code{\link{kmdoubleequal}()},
\code{\link{kmsymmsetdiff}()}
}
\concept{Utilities}
