% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmsf2basis.R
\name{kmsf2basis}
\alias{kmsf2basis}
\title{Derive a basis from a surmise function}
\usage{
kmsf2basis(sf)
}
\arguments{
\item{sf}{Surmise function}
}
\value{
Matrix representing the basis.
}
\description{
\code{kmsf2basis} expects a surmise function data frame
and returns the corresponding basis.
}
\seealso{
Other Different representations for knowledge spaces: 
\code{\link{kmbasis}()},
\code{\link{kmsurmisefunction}()},
\code{\link{kmsurmiserelation}()},
\code{\link{kmunionclosure}()}
}
\concept{Different representations for knowledge spaces}
