% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmspace.R
\name{kmspace}
\alias{kmspace}
\title{Convert a binary matrix to a \code{kmspace} object}
\usage{
kmspace(x)
}
\arguments{
\item{x}{Binary matrix representing a family of sets}
}
\value{
\code{kmspace} object
}
\description{
\code{kmspace()} returns a \code{kmspace} object for a binary matrix.
}
\examples{
m <- as.matrix(c(1,0,0,0,1,0,1,1,1), nrow=3, byrow=TRUE)
kmspace(m)

}
\seealso{
Other Constructors: 
\code{\link{kmfamset}()},
\code{\link{kmstructure}()}
}
\concept{Constructors}
