% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmSR2basis.R
\name{kmSR2basis}
\alias{kmSR2basis}
\title{Determine the basis of a knowledge space from a surmise relation}
\usage{
kmSR2basis(sr)
}
\arguments{
\item{sr}{Surmise relation}
}
\value{
Basis
}
\description{
\code{kmSR2basis} takes a surmise relation and returns the
corresponding basis.
}
\seealso{
Other Different representations for knowledge spaces: 
\code{\link{kmbasis.kmsurmiserelation}()},
\code{\link{kmbasis.matrix}()},
\code{\link{kmsf2basis}()},
\code{\link{kmsurmisefunction}()},
\code{\link{kmsurmiserelation}()},
\code{\link{kmunionclosure}()}
}
\concept{Different representations for knowledge spaces}
