% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmSRdiagram.R
\name{kmSRdiagram}
\alias{kmSRdiagram}
\title{Plot the Hasse diagram of a basis stored as a matrix}
\usage{
kmSRdiagram(structure, horizontal = TRUE, colors = NULL)
}
\arguments{
\item{structure}{Binary matrix representing a surmise relation}

\item{horizontal}{Boolean defining orientation of the graph, default TRUE}

\item{colors}{Color vector (default NULL)}
}
\description{
\code{kmSRdiagram} takes a matrix representing a surmise relation
and a color vector and draws a Hasse diagram. If the color vector is
NULL the states are drawn in green.
}
\seealso{
Other Plotting knowledge structures: 
\code{\link{kmbasisdiagram}()},
\code{\link{kmhasse}()}
}
\concept{Plotting knowledge structures}
\keyword{math}
