% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmsrvalidate.R
\name{kmSRvalidate}
\alias{kmSRvalidate}
\title{Validate a surmise relation against a data set}
\usage{
kmSRvalidate(data, sr)
}
\arguments{
\item{data}{Binary matrix representing a set of response patterns}

\item{sr}{Binary matrix representing a surmise relation}
}
\value{
A list with two elements:
\describe{
\item{gamma}{Goodman & Kruskal's gamma index}
\item{VC}{Violational Coefficient}
}
}
\description{
\code{kmSRvalidate} returns a list with two elements,
Goodman & Kruskal's gamma value and the violational
coefficient (VC).
}
\examples{
kmSRvalidate(xpl$data, xpl$sr)

}
\seealso{
Other Validating knowledge spaces: 
\code{\link{kmdist}()},
\code{\link{kmvalidate}()}
}
\concept{Validating knowledge spaces}
