% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmstructure.R
\name{kmstructure}
\alias{kmstructure}
\title{Convert a binary matrix to a \code{kmstructure} object}
\usage{
kmstructure(x)
}
\arguments{
\item{x}{Binary matrix representing a family of sets}
}
\value{
\code{kmstructure} object
}
\description{
\code{kmstructure()} returns a \code{kmstructure} object after checking that
the passed object is a binary matrix without double rows. The empty set
and the full item set are added if missing.
}
\examples{
m <- as.matrix(c(1,0,0,0,1,0,1,1,1), nrow=3, byrow=TRUE)
kmstructure(m)

}
\seealso{
Other Constructors: 
\code{\link{kmfamset}()},
\code{\link{kmspace}()}
}
\concept{Constructors}
