% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmsurmisefunction.R
\name{kmsurmisefunction}
\alias{kmsurmisefunction}
\title{Compute the surmise function for a knowledge space or basis}
\usage{
kmsurmisefunction(x)
}
\arguments{
\item{x}{Binary matrix representing a knowledge space or basis}
}
\value{
Data frame  representing the surmise unction of \code{x}.
}
\description{
\code{kmsurmisefunction} returns a data frame representing the
surmise function for a knowledge space or basis. The rows of the
data frame are ordered by item name.
}
\examples{
kmsurmisefunction(xpl$space)

}
\seealso{
Other Different representations for knowledge spaces: 
\code{\link{kmbasis}()},
\code{\link{kmsurmiserelation}()},
\code{\link{kmunionclosure}()}
}
\concept{Different representations for knowledge spaces}
\keyword{math}
