% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmsymmsetdiff.R
\name{kmsymmsetdiff}
\alias{kmsymmsetdiff}
\alias{kmsetdistance}
\title{Compute the symmetric set difference between two sets}
\usage{
kmsymmsetdiff(x, y)

kmsetdistance(x, y)
}
\arguments{
\item{x}{Binary vector representing a set}

\item{y}{Binary vector representing a set}
}
\value{
kmsymmsetdiff: Symmetric set difference between 'x' and 'y'

kmsetdistance: Distance between the sets 'x' and 'y', i.e. the cardinality
of the symmetric set difference
}
\description{
Compute the symmetric set difference between two sets
}
\examples{
kmsymmsetdiff(c(1,0,0), c(1,1,0))

kmsetdistance(c(1,0,0), c(1,1,0))

}
\keyword{math}
