% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmtrivial.R
\name{kmtrivial}
\alias{kmtrivial}
\alias{kmminimalspace}
\alias{kmmaximalspace}
\title{Create trivial knowledge spaces}
\usage{
kmminimalspace(noi)

kmmaximalspace(noi)
}
\arguments{
\item{noi}{Number of items}
}
\value{
A binary matrix representing the respective knowledge space
}
\description{
These functions create trivial knowledge spaces of a given item number. The minimal
space contains just the empty set and the full item set while the maximal space is
equal to the power set.
}
\details{
Please note that the computation time for creating large power sets can grow quite
large easily.
}
\examples{
kmminimalspace(5)
kmmaximalspace(5)

}
\concept{Trivial knowledge spaces}
\keyword{math}
