% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsg.R
\docType{data}
\name{phsg}
\alias{phsg}
\title{Knowledge space on linear functions}
\format{
A list containing the basis and the list of items
}
\usage{
phsg
}
\description{
Basis of a small knowledge space and list of items on linear functions
used in a manuscript by Steiner et al.
}
\seealso{
Other Data: 
\code{\link{cad}},
\code{\link{fractions}},
\code{\link{readwrite}},
\code{\link{xpl}}
}
\concept{Data}
\keyword{datasets}
