% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpl.R
\docType{data}
\name{xpl}
\alias{xpl}
\title{Small example knowledge space}
\format{
A list containing the basis, the space, the surmise relation,
        the surmise function, and the two data matrices data and sim.
}
\usage{
xpl
}
\description{
Basis and space matrix, surmise relation and surmise function of a small
fictional knowledge space, and two data sets (data (7 patterns) and
sim (500 patterns)to be used in examples. The latter was produced from
the space with \code{kmsimulate()} with beta and eta values of 0.1.
}
\seealso{
Other Data: 
\code{\link{cad}},
\code{\link{fractions}},
\code{\link{readwrite}}
}
\concept{Data}
\keyword{data}
