% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmbasis.R
\name{kmbasis}
\alias{kmbasis}
\title{Generic \code{kmbasis()} function}
\usage{
kmbasis(x, ...)
}
\arguments{
\item{x}{Family of sets or surmise relation in matrix representation}

\item{...}{Optional additional parameters}
}
\value{
Basis for a knowledge space
}
\description{
Generic \code{kmbasis()} function
}
