% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmfringe.R
\name{kmfringe}
\alias{kmfringe}
\title{Compute the fringe of a state within a knowledge structure}
\usage{
kmfringe(state, struct)
}
\arguments{
\item{state}{Binary vector representing a knowledge state}

\item{struct}{Binary matrix representing a knowledge structure}
}
\value{
Binary vector representing the fringe
}
\description{
\code{kmfringe} computes the fringe of a state
within a knowledge structure, i.e. the set of items by which
the state differs from its neighbours.
}
\examples{
kmfringe(c(1,0,0,0), xpl$space)

}
\seealso{
Other Neighbourhood & fringe: 
\code{\link{kmneighbourhood}()}
}
\concept{Neighbourhood & fringe}
