% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmminimalfamset.R
\name{kmminimalfamset}
\alias{kmminimalfamset}
\title{Determine subfamily of minimal states in a \code{kmfamset}}
\usage{
kmminimalfamset(x)
}
\arguments{
\item{x}{kmfamset}
}
\value{
kmfamset of minimal states

\code{kmminimalfamset()} determines the sub family of minimal sets
within a famset. If \code{x} is a knowledge structure, the result will
be a family containing (only) the empty set because the empty set is a
state in any knowledge structure and is its minimal state.

The concept of the minimal state of a family of sets should not be confused
with the concept of knowledge states minimal for some item.
}
\description{
Determine subfamily of minimal states in a \code{kmfamset}
}
\examples{
m <- matrix(c(1,0,0,0,1,0,1,0,1), ncol=3, byrow=TRUE)
m
kmminimalfamset(kmfamset(m))


}
\seealso{
Other Utilities:
\code{\link[=kmcolors]{kmcolors()}},
\code{\link[=kmdoubleequal]{kmdoubleequal()}},
\code{\link[=kmsetiselement]{kmsetiselement()}},
\code{\link[=kmsymmsetdiff]{kmsymmsetdiff()}}
}
\concept{Utilities}
