% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmneighbourhood.R
\name{kmneighbourhood}
\alias{kmneighbourhood}
\title{Compute the neighbourhod of a state within a knowledge structure}
\usage{
kmneighbourhood(state, struct)
}
\arguments{
\item{state}{Binary vector representing a knowledge state}

\item{struct}{Binary matrix representing a knowledge structure}
}
\value{
Matrix containing the neighbouring states, one per row
}
\description{
\code{kmneighbourhood} computes the neighbourhood of a state
within a knowledge structure, i.e. the family of all other
states with a symmetric set difference of 1.
}
\examples{
kmneighbourhood(c(1,1,0,0), xpl$space)

}
\seealso{
Other Neighbourhood & fringe: 
\code{\link{kmfringe}()}
}
\concept{Neighbourhood & fringe}
\keyword{math}
