% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmSF2basis.R
\name{kmSF2basis}
\alias{kmSF2basis}
\title{Derive a basis from a surmise function}
\usage{
kmSF2basis(sf)
}
\arguments{
\item{sf}{Surmise function}
}
\value{
Matrix representing the basis.
}
\description{
\code{kmSF2basis} expects a surmise function data frame
and returns the corresponding basis.
}
\seealso{
Other Different representations for knowledge spaces: 
\code{\link{kmSR2basis}()},
\code{\link{kmbasis.kmsurmisefunction}()},
\code{\link{kmbasis.kmsurmiserelation}()},
\code{\link{kmbasis.matrix}()},
\code{\link{kmsurmisefunction}()},
\code{\link{kmsurmiserelation}()},
\code{\link{kmunionclosure}()}
}
\concept{Different representations for knowledge spaces}
