% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xpl.R
\docType{data}
\name{xpl}
\alias{xpl}
\title{Small example knowledge spaces}
\format{A list containing the basis, the space, and the data matrix}
\usage{
xpl
}
\description{
Basis and space matrix of a small fictional knowledge space, and a small data set to be
used in examples.
}
\keyword{data}
